+++ src/main/java/dev/prvt/yawiki/core/wikititle/cache/application/WikiPageTitleExistenceCheckerImpl.java
@@ -1,8 +1,8 @@
-package dev.prvt.yawiki.core.wikititle.localcache.application;
+package dev.prvt.yawiki.core.wikititle.cache.application;
 
 import dev.prvt.yawiki.core.wikipage.domain.model.WikiPageTitle;
 import dev.prvt.yawiki.core.wikititle.existence.WikiPageTitleExistenceChecker;
-import dev.prvt.yawiki.core.wikititle.localcache.domain.CacheStorage;
+import dev.prvt.yawiki.core.wikititle.cache.domain.CacheStorage;
 import java.util.Collection;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
