+++ src/main/java/dev/prvt/yawiki/core/wikititle/cache/domain/updater/RemoteChangesRepository.java
@@ -1,4 +1,4 @@
-package dev.prvt.yawiki.core.wikititle.localcache.domain.updater;
+package dev.prvt.yawiki.core.wikititle.cache.domain.updater;
 
 import java.time.LocalDateTime;
 import java.util.List;
