+++ src/main/java/dev/prvt/yawiki/core/wikititle/cache/exception/CacheNotInitializedException.java
@@ -1,4 +1,4 @@
-package dev.prvt.yawiki.core.wikititle.localcache.exception;
+package dev.prvt.yawiki.core.wikititle.cache.exception;
 
 public class CacheNotInitializedException extends IllegalStateException {
