+++ src/test/java/dev/prvt/yawiki/core/wikititle/localcache/infra/initializer/InitialCacheDataReaderImplTest.java
@@ -3,8 +3,9 @@ package dev.prvt.yawiki.core.wikititle.localcache.infra.initializer;
 import dev.prvt.yawiki.core.wikipage.domain.model.WikiPage;
 import dev.prvt.yawiki.core.wikipage.domain.model.WikiPageFactory;
 import dev.prvt.yawiki.core.wikipage.domain.model.WikiPageTitle;
-import dev.prvt.yawiki.core.wikititle.localcache.domain.InitialCacheData;
-import dev.prvt.yawiki.core.wikititle.localcache.domain.initializer.InitialCacheDataReader;
+import dev.prvt.yawiki.core.wikititle.cache.domain.InitialCacheData;
+import dev.prvt.yawiki.core.wikititle.cache.domain.initializer.InitialCacheDataReader;
+import dev.prvt.yawiki.core.wikititle.cache.infra.initializer.InitialCacheDataReaderImpl;
 import dev.prvt.yawiki.fixture.WikiPageFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
